Separated player and ship tests
