+++ core/src/test/java/com/mygdx/game/ShipTests.java
@@ -86,33 +86,6 @@ public class ShipTests {
 	}
 
 	@Test
-	public void playerProgressTasks() {
-		Player player = new Player();
-
-		Vector2 questLocA = new Vector2(100,100);
-		Vector2 questLocB = new Vector2(player.getPosition());
-
-		LocateQuest questA = new LocateQuest(questLocA,1);
-		LocateQuest questB = new LocateQuest(questLocB,1);
-
-		boolean questStatusA = questA.checkCompleted(player);
-		boolean questStatusB = questB.checkCompleted(player);
-
-		assertFalse("Locate Quest shown as complete when isn't",questStatusA);
-		assertTrue("Locate Quest shown as incomplete when isn't",questStatusB);
-
-		College collegeToKill = new College(1);
-		KillQuest killCollege = new KillQuest(collegeToKill);
-		boolean questStatusC = killCollege.checkCompleted(player);
-
-		collegeToKill.killThisCollege(null);
-		boolean questStatusD = killCollege.checkCompleted(player);
-
-		assertFalse("Kill Quest shown as complete when isn't",questStatusC);
-		assertTrue("Kill Quest shown as incomplete when isn't",questStatusD);
-	}
-
-	@Test
 	public void shipFires() {
 		Ship ship = new Ship();
 		CannonBall cannonBall = GameManager.getCurrentCannon();
@@ -124,44 +97,5 @@ public class ShipTests {
 		ship.shoot(shootDirection);
 		Vector2 cannonNewPos = cannonT.getPosition().cpy();
 		assertNotEquals("Cannonball has not been fired", cannonStartPos, cannonNewPos);
-	}
-
-	@Test
-	public void playerMoney() {
-		// player
-		GameManager.CreatePlayer();
-		Player player = GameManager.getPlayer();
-		int initial = player.getPlunder();
-		Vector2 playerLoc = new Vector2(player.getPosition());
-		//quests
-		QuestManager.Initialize();
-		assertEquals("Player started with more than 0 loot", 0, initial);
-		College collegeToKill = new College(1);
-		KillQuest killCollege = new KillQuest(collegeToKill);
-		LocateQuest locateQuest = new LocateQuest(playerLoc, 1);
-
-		QuestManager.addQuest(killCollege);
-
-		collegeToKill.killThisCollege(null);
-		QuestManager.checkCompleted();
-
-		int after = player.getPlunder();
-		assertTrue("Player hasn't gained loot after completing killCollege task", (after > initial));
-
-
-		initial = after; //new initial after 1 complete quest
-
-		QuestManager.addQuest(locateQuest);
-		QuestManager.checkCompleted();
-
-		after = player.getPlunder();
-		assertTrue("Player hasn't gained loot after picking up loot chest ", (after > initial));
-		//TODO: Check loot on enemy kill
-
-	}
-
-	@Test
-	public void playerPoints() {
-
 	}
 }
